Adding the file that does positions predictions
The top of the file is the relevant code
The bottom of the file is code used for testing
that can be ignored for now (it's labeled as "test code").
This file creates the predictions results that can be used to
compure ROC/PR curves, and also the AUC scores for refrence
with the ones computed by R code.